+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 3.3 (unreleased)
 ----------------
 
+- Add support for bytes and matrix multiplication.
+
 - Remove the Python 2-only C extension.
 
 3.2 (2017-04-26)